scraper-api/requirements.txt
fastapi
uvicorn
python-jobspy
pandas
pydantic
python-dotenv
anthropic
